requirements.txt
streamlit
openai
python-dotenv
PyPDF2
llama-index-core
chromadb